+++ SpendWisely/Models/Income.cs
@@ -24,7 +24,7 @@ namespace SpendWisely.Models
         [Required(ErrorMessage = "El monto del ingreso es obligatorio.")]
         [Display(Name ="Monto")]
         [DataType(DataType.Currency)]
-        public decimal amount { get; set; }
+        public int amount { get; set; }
 
         public int userId { get; set; }
